backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
setuptools>=65.0.0
pandas>=1.5.0,<2.0.0
numpy>=1.24.0,<1.26.0
scikit-learn==1.3.2
xgboost==2.0.3
shap==0.43.0
lime==0.2.0.1
# pycaret==3.0.4  # Optional - commented out due to dependency conflicts
mlflow==2.8.1
openpyxl==3.1.2
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
plotly==5.18.0
seaborn==0.13.0
matplotlib==3.8.2
scipy==1.11.4
fairlearn==0.9.0
aiofiles==23.2.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
kaleido==0.2.1
python-docx==1.1.0
reportlab==4.0.9
# PostgreSQL support
psycopg2-binary==2.9.9
# Cloud storage (optional - install as needed)
# boto3>=1.28.0  # For AWS S3 - uncomment if using S3
# vercel-blob>=0.1.0  # For Vercel Blob - uncomment if using Vercel Blob
gunicorn==21.2.0

